+++ design-system-docs/server/roda_app.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Roda is a simple Rack-based framework with a flexible architecture based
 # on the concept of a routing tree. Bridgetown uses it for its development
 # server, but you can also run it in production for fast, dynamic applications.
@@ -14,7 +16,7 @@ class RodaApp < Bridgetown::Rack::Roda
   # plugin :bridgetown_ssr
   # plugin :bridgetown_routes
 
-  route do |r|
+  route do |_r|
     # Load all files in server/routes and src/_routes if "bridgetown-routes" is loaded
     # (see server/routes/hello.rb.sample)
     Bridgetown::Rack::Routes.start! self