+++ testing/features/support/components/breadcrumbs/no_collapse.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Breadcrumbs
-  class NoCollapse < Default
-    set_url "/iframe.html?id=components-breadcrumbs--not-collapsing&viewMode=story"
-  end
-end