+++ testing/features/support/components/targeted_content/fallback.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module TargetedContent
-  class Fallback < Default
-    set_url "/iframe.html?id=components-targeted-content--fallback&viewMode=story"
-  end
-end