+++ testing/features/support/forms/input/error_message.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Input
-  class ErrorMessage < Basic
-    set_url "/iframe.html?id=forms-input--with-error&viewMode=story"
-  end
-end