+++ testing/visual-regression/backstop-config.js
@@ -1,3 +1,0 @@
-const backstopConfigCommon = require('./backstop-common');
-
-module.exports = backstopConfigCommon('http://host.docker.internal:6006');